+++ lib/src/registry_api/component_description.dart
@@ -1,24 +1,13 @@
-import 'package:duit_kernel/src/extensions.dart';
-
-///Description of the component for registering it
-///in the [DuitRegistry] under the key corresponding to the [tag] property
+/// Description of the component for registering it
+/// in the [DuitRegistry] under the key corresponding to the [tag] property
 final class DuitComponentDescription {
   final String tag;
   final Map<String, dynamic> data;
-  final Iterable<String> referencePaths;
 
   DuitComponentDescription({
     required this.tag,
     required this.data,
-    required this.referencePaths,
   });
-
-  static Iterable<String> _extractRefPaths(Map<String, dynamic> json) {
-    return json
-        .flatMap(breakPointKey: "refs")
-        .keys
-        .where((key) => key.endsWith("refs"));
-  }
 
   factory DuitComponentDescription.fromJson(Map<String, dynamic> json) {
     assert(
@@ -29,7 +18,6 @@ final class DuitComponentDescription {
     return DuitComponentDescription(
       tag: json['tag'],
       data: root,
-      referencePaths: [],
     );
   }
 }