+++ lib/src/registry_api/index.dart
@@ -1,0 +1,6 @@
+export 'registry.dart';
+export 'value_reference.dart';
+export 'component_description.dart';
+export 'model_factory.dart';
+export 'attributes_factory.dart';
+export 'build_factory.dart';